Fixed inhitance
Also removed usless files, cleaned up some images and adjusted start position

+++ Source/Game.h
@@ -37,8 +37,8 @@ private:
 	bool exit = false;                 /**< Exit boolean. If true the game loop will exit. */
 	ASGE::Sprite* sprite = nullptr;    /**< Sprite Object. The background sprite. */
 
-	int count = 0;
-	int game_speed = 30;
+	float count = 0.0f;
+	float game_speed = 30;
 	int new_pos[1];
 	int menu_option = 1;
